Move MockWebServer setup to new file
This new file is shared across both unittests and espresso tests
