+++ JTime-android/build.gradle
@@ -33,6 +33,8 @@ android {
     }
     sourceSets {
         main.java.srcDirs += 'src/main/kotlin'
+        test.java.srcDirs += 'src/sharedTest/kotlin'
+        androidTest.java.srcDirs += 'src/sharedTest/kotlin'
     }
 
     signingConfigs {
